Cache found files in CoverageStartIncluder and add tests
Since the coverage starter files should be included in every request,
it is not efficient to search for the files to be included in every
request. This commit caches the found files' paths. By this way, there
is only one search performed.

The files are not included in the constructor method now. They are
included when `includeFiles()` method is called. Hence, example file for
CoverageStarterIncluder is updated as well.

This commit also adds `tests` directory. PHPUnit is added as a dev
dependency. Tests for CoverageStartIncluder is added as well.