Add tests/site, use an Apache image, use Xdebug 2.8.1 +
+ Moves env/php directory to env/services/php

This site is attached to the PHP-Apache container so that we can produce
coverage dumps and reports to test if the tools work properly. An
external NGINX container should be up and running. It should be served
under "nginx" network, which is an external network that can be created
by running "docker network create nginx".